+++ database/migrations/2022_11_18_063857_create_hardware_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
         Schema::create('hardware', function (Blueprint $table) {
             $table->id();
             $table->string('code');
-            $table->tinyInteger('hardware_category_id');
+            $table->bigInteger('hardware_category_id');
             $table->string('name');
-            $table->tinyInteger('manufacturer_id');
+            $table->bigInteger('manufacturer_id');
             $table->string('serial_number');
             $table->string('status');
             $table->date('warranty_start')->nullable();
@@ -28,12 +28,12 @@ return new class extends Migration
             $table->string('image_format')->default('jpg');
             $table->text('remark')->nullable();
             $table->string('service_code')->nullable();
-            $table->tinyInteger('hardware_type_id');
-            $table->tinyInteger('hardware_model_id');
-            $table->tinyInteger('processor_id');
-            $table->tinyInteger('memory_id');
-            $table->tinyInteger('graphic_card_id');
-            $table->tinyInteger('storage_id');
+            $table->bigInteger('hardware_type_id');
+            $table->bigInteger('hardware_model_id');
+            $table->bigInteger('processor_id');
+            $table->bigInteger('memory_id');
+            $table->bigInteger('graphic_card_id');
+            $table->bigInteger('storage_id');
             $table->string('computer_name')->nullable();
             $table->timestamps();
         });
